fix(routes): routes hotfix [Delivers #167474497]
